Add `-check` flag to check if local CA is installed
When using `mkcert` as part of an automation script you sometimes want to know if the user already installed the local CA or not.

Running `mkcert -install` out of the blue could confuse users because they don't know what's happening.

Therefore, it's better to first check if the local CA is installed with `mkcert -check`. If that returns a zero exit code, all is good. If it returns a non-zero exit code, it means that the local CA is not installed and the user can be instructed on what to do next (`mkcert -install`).

## How does it look?

```
$ mkcert -check
Note: the local CA is not installed in the system trust store.
Note: the local CA is not installed in the Firefox trust store.
Run "mkcert -install" for certificates to be trusted automatically ⚠️

$ echo $?
1

$ mkcert -install
The local CA is now installed in the system trust store! ⚡️
The local CA is now installed in the Firefox trust store (requires browser restart)! 🦊

$ mkcert -check
All good!

$ echo $?
0
```